messages: make Request an enum, introduce one for the Response result
Having enums allows for a simpler generalistic handling of message on
the server side. The Response keeps the flexibility of being passed an
enum for servers' callbacks, while being able to expect a specific message
on client side.

The Request is interstingly using both an enum for itself in order to
pair the method name with the parameters struct, and an additional one
is present only for the params in order to not bother with the method's
str lifetime in callbacks (or worse, callbacks returning futures). It
will also strip off the id once we have it too.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>